+++ src/_helpers.ts
@@ -291,7 +291,11 @@ export function mkRelativePath(absRoot: string, absPath: string): string {
 }
 
 export function mkRelativePathReproducibleHash(absRoot: string, absPath: string): string {
-  return sha256hex(mkRelativePath(absRoot, absPath).replaceAll(sep, '/'))
+  return sha256hex(mkPosixPathLike(mkRelativePath(absRoot, absPath)))
+}
+
+export function mkPosixPathLike(p: string): string {
+  return p.replaceAll(sep, '/')
 }
 
 // endregion relative paths