+++ HISTORY.md
@@ -7,12 +7,15 @@ All notable changes to this project will be documented in this file.
 <!-- add unreleased items here -->
 
 * Added
+  * SBOM result now includes a dependency graph ([#11] via [#161])
   * SBOM result now includes components’ `scope` (via [#143])
   * SBOM result may now include excluded, tree‑shaken components ([#142] via [#143])  
     The respective component’s `scope` is properly set to `"excluded"`.
 
+[#11]: https://github.com/CycloneDX/cyclonedx-esbuild/issues/11
 [#142]: https://github.com/CycloneDX/cyclonedx-esbuild/issues/142
 [#143]: https://github.com/CycloneDX/cyclonedx-esbuild/pull/143
+[#161]: https://github.com/CycloneDX/cyclonedx-esbuild/pull/161
 
 ## 1.3.1 - 2026-03-18