+++ HISTORY.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file.
     The respective component’s `scope` is properly set to `"excluded"`.
   * Components might have properties following [`cdx:esbuild` Namespace Taxonomy](https://cyclonedx.github.io/cyclonedx-property-taxonomy/cdx/esbuild.html)
     and [`cdx` Namespace Taxonomy](https://cyclonedx.github.io/cyclonedx-property-taxonomy/cdx.html)
-    .
+    . (via [#143])
 
 [#142]: https://github.com/CycloneDX/cyclonedx-esbuild/issues/142
 [#143]: https://github.com/CycloneDX/cyclonedx-esbuild/pull/143